feat: gen 自动生成 gorm 模型 - site category

+++ internal/api/site/func_create.go
@@ -10,13 +10,13 @@ import (
 )
 
 type createRequest struct {
-	CategoryId int32  `form:"category_id"`
+	CategoryId int64  `form:"category_id"`
 	Url        string `form:"url"`
 }
 
 type createResponse struct {
-	SuccessCount int32 `json:"successCount"`
-	FailCount    int32 `json:"failCount"`
+	SuccessCount int64 `json:"successCount"`
+	FailCount    int64 `json:"failCount"`
 }
 
 // Create 创建网站
